% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{collapseInHtml}
\alias{collapseInHtml}
\title{Function to create collapse with button in html output}
\usage{
collapseInHtml(input, id, titleButton, type = c("function", "text"),
  color = "#479ace", borderColor = "#479ace")
}
\arguments{
\item{input}{the text or function (see type) that should be printed}

\item{id}{the id of the environment}

\item{titleButton}{the title of the button}

\item{type}{whether the input is a function or text}

\item{color}{the color of the text in the button}

\item{borderColor}{the color of the border of the button}
}
\value{
html code with input in a collapsible environment
}
\description{
Function to create collapse with button in html output
}
\author{
Kirsten Van Hoorde
}
