% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPathStartTemplate}
\alias{getPathStartTemplate}
\title{get path of start template contained in the \link{exampleRTemplatePackage} package}
\usage{
getPathStartTemplate(template)
}
\arguments{
\item{template}{string, name of the template file}
}
\value{
path of start template, from the installed \link{exampleRTemplatePackage} library
}
\description{
get path of start template contained in the \link{exampleRTemplatePackage} package
}
\author{
Laure Cougnaud and Kirsten Van Hoorde
}
