% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPathTemplate}
\alias{getPathTemplate}
\title{get path of template}
\usage{
getPathTemplate(template, userType = c("user", "developer"))
}
\arguments{
\item{template}{string, name of the template file}

\item{userType}{string, use type, either:
\itemize{
\item{'user': }{the template path is extracted from the installed \link{exampleRTemplatePackage} library}
\item{'developer:' }{the name of the template is returned, assuming that the developer works
in the 'inst/templates' folder of the package directly
}
}}
}
\value{
string with output path(s) of the specified file
}
\description{
get path of template
}
\author{
Laure Cougnaud and Kirsten Van Hoorde
}
