% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPrefixSection}
\alias{getPrefixSection}
\title{get prefix for section in markdown ('#') for a number of level}
\usage{
getPrefixSection(level)
}
\arguments{
\item{level}{integer, level of section, e.g. 2}
}
\value{
string with prefix for section, e.g. '##'
}
\description{
get prefix for section in markdown ('#') for a number of level
}
\author{
Laure Cougnaud
}
