% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisonPlot.R
\name{pairwiseComparisonPlot}
\alias{pairwiseComparisonPlot}
\title{wrapper to generate pairwise comparison plot}
\usage{
pairwiseComparisonPlot(dataset, variables, description, colorVariable = NULL,
  colorPalette = "black", typePlot = c("baseR", "ggplot2", "plotly",
  "rbokeh"))
}
\arguments{
\item{dataset}{data.frame with data}

\item{variables}{columns of \code{dataset} with variable for x and y}

\item{description}{text description of the pairwise comparison (used for the title)}

\item{colorVariable}{column of \code{dataset} with color variable}

\item{colorPalette}{string with color palette, if \code{colorVariable}
\itemize{
    \item{is specified: }{named string with color for each level (name) of \code{colorVariable}}
 \item{is not specified: }{unique element}
}}

\item{typePlot}{string with type of the plot, either 'baseR' (\code{plot}),
'ggplot2', 'plotly', 'rbokeh'}
}
\value{
no returned value, a plot is drawn in the current window
}
\description{
wrapper to generate pairwise comparison plot
}
\author{
Laure Cougnaud
}
