% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPathStartTemplate}
\alias{getPathStartTemplate}
\title{get path of start template contained in the \code{exampleRTemplatePackage} package}
\usage{
getPathStartTemplate()
}
\value{
path of start template, from the installed \code{exampleRTemplatePackage} library
}
\description{
get path of start template contained in the \code{exampleRTemplatePackage} package
}
\author{
Laure Cougnaud and Kirsten Van Hoorde
}
